requirements.txt
redis
redis-om
requests
sentry-sdk
python-telegram-bot
